requirements.txt
owslib==0.23.0
requests==2.25.1
pandas==1.0.3
geopandas==0.8.1
Shapely==1.6.4.post2
pytest==6.2.2
pytest-cov==2.11.1
psycopg2==2.8.4
python-dotenv==0.12.0